+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## 0.3.40
 
 * Relax `fast-logger` upper bound from 3.2 to 3.3
+* Add `Alternative` instances for `LoggingT` and `NoLoggingT`
 
 ## 0.3.39